Updated and tested Gestures few Tweaks for the finger detection touch collection

+++ IntelPCSDK-Manager/Input/IController.cs
@@ -7,7 +7,7 @@ namespace IntelPCSDK_Manager.Input
         void Recognise(object input, bool debug = false);
         void DebugDraw(Microsoft.Xna.Framework.Graphics.SpriteBatch spritebatch);
         void EnableDebug(Microsoft.Xna.Framework.Graphics.GraphicsDevice device);
-        void SetScale(int scale);
+        void SetScale(double scale);
 
     }
 }
